+++ Shopping/src/Vue/VueProduitsBoissons.java
@@ -100,7 +100,7 @@ public class VueProduitsBoissons extends JPanel {
                     JPanel productPanel = ProductPanelFactory.createProductPanel(
                             article.getNomArticle(),         // Nom du produit
                             article.getCategorieArticle(),   // Description ou catégorie
-                            (Double)(double) article.getPrixArticle(),        // Prix
+                            article.getPrixArticle(),        // Prix
                             article.getNomArticle()        // Nom du porduit
                     );
